+++ cli/src/main/kotlin/com/sermilion/personalgraph/cli/di/CliComponent.kt
@@ -3,6 +3,7 @@ package com.sermilion.personalgraph.cli.di
 import com.sermilion.personalgraph.common.di.AppScope
 import com.sermilion.personalgraph.common.di.CoreComponent
 import com.sermilion.personalgraph.data.di.DataComponent
+import com.sermilion.personalgraph.domain.repository.ConsolidationService
 import com.sermilion.personalgraph.domain.scaffold.VaultScaffolder
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Provides
@@ -14,4 +15,5 @@ abstract class CliComponent(@get:Provides val vaultRoot: Path) :
   CoreComponent,
   DataComponent {
   abstract val vaultScaffolder: VaultScaffolder
+  abstract val consolidationService: ConsolidationService
 }